fix(fga): tune batch-check parallelism, add tests
Raise batchCheckMaxParallelRequests from 1 to 4 so large batches
(query-service's MaxPageSize of 1000, chunked at 50/request) don't
risk exceeding fgaHTTPTimeout/accessCheckHandlerTimeout by fully
serializing. Type the constant as int32 to drop a conversion, and
add tests covering the fan-out bound and the handler deadline.

Issue: #2957

+++ fga_client.go
@@ -39,7 +39,7 @@ func (c FgaAdapter) BatchCheck(
 	ctx context.Context,
 	request ClientBatchCheckRequest,
 ) (*openfga.BatchCheckResponse, error) {
-	maxParallelRequests := int32(batchCheckMaxParallelRequests)
+	maxParallelRequests := batchCheckMaxParallelRequests
 	return c.OpenFgaClient.BatchCheck(ctx).
 		Body(request).
 		Options(BatchCheckOptions{MaxParallelRequests: &maxParallelRequests}).
